requirements.txt
opencv-python~=4.10.0.84
face-recognition~=1.3.0
numpy~=1.26.4
pytesseract~=0.3.13
streamlit~=1.39.0
pillow~=10.4.0